#11278: Compiling TTNN with GCC-12 (#11482)
* #0: install g++-12 into ubuntu 22.04 image

* #0: test g++-12 and clang-17 build on 22.04

* #0: fix g++-12 compile error

* #0: convert row col type to size_t from int

* #0: build release variants on 22.04 as well

* #11278: Compile TTNN with GCC-12

* #11278: remove a few build variants given our GH runner capacity constraint

* #0: fix bad rebase

* #0: disable test in python

* #0: install only required dependencies as we're out of disk space when building on GH runners

* #0: add steps to check disk space during compile on GH runners

* #0: re-add pandocs as a dependency, needed by docs build

* #0: Run relwithdebinfo build on self-hosted runners

* #0: clean up

---------

+++ tests/tt_metal/tt_metal/unit_tests/compute/test_dropout_sfpu_compute.cpp
@@ -250,8 +250,8 @@ TEST_F(DeviceFixture, ComputeDropout) {
         unit_tests::compute::sfpu::dropout::DropoutConfig test_config = {
 		.probability = probability,
 		.fill_constant = fill_constant,
-		.seed_0 = rand(),
-		.seed_1 = rand()
+		.seed_0 = static_cast<uint32_t>(rand()),
+		.seed_1 = static_cast<uint32_t>(rand())
         };
         unit_tests::compute::sfpu::dropout::test_dropout(this->devices_.at(0), test_config);
     }
